enhanced: 기존 인게임 정보 조회, 티어 기준 랭킹 정보 조회 기능 개선

+++ src/main/java/onlysolorank/apiserver/api/service/dto/SummonerPlayDto.java
@@ -30,7 +30,7 @@ public class SummonerPlayDto {
     private Integer totalWin;
     private Integer totalDefeat;
     private boolean isPerfect = false;
-    private Integer championId;
+    private Long championId;
     private String championName;
 
     public SummonerPlayDto(SummonerPlay summonerPlay) {
